文献求助添加渠道限制 统计接口添加渠道参数

+++ src/main/java/com/wd/cloud/docdelivery/pojo/entity/Permission.java
@@ -28,7 +28,10 @@ public class Permission extends AbstractEntity {
 
     private String orgName;
 
-    private Integer channel;
+    /**
+     * 渠道ID
+     */
+    private Long channel;
     /**
      * 校内：1，登陆：2，验证：4，最后相加得到权限
      */
